+++ lib/src/backend.rs
@@ -424,26 +424,6 @@ impl Tree {
     }
 }
 
-/// Calculates common prefix length of two bytes. The length to be returned is
-/// a number of hexadecimal digits.
-pub fn common_hex_len(bytes_a: &[u8], bytes_b: &[u8]) -> usize {
-    iter_half_bytes(bytes_a)
-        .zip(iter_half_bytes(bytes_b))
-        .take_while(|(a, b)| a == b)
-        .count()
-}
-
-fn iter_half_bytes(bytes: &[u8]) -> impl ExactSizeIterator<Item = u8> + '_ {
-    (0..bytes.len() * 2).map(|i| {
-        let v = bytes[i / 2];
-        if i & 1 == 0 {
-            v >> 4
-        } else {
-            v & 0xf
-        }
-    })
-}
-
 pub fn make_root_commit(root_change_id: ChangeId, empty_tree_id: TreeId) -> Commit {
     let timestamp = Timestamp {
         timestamp: MillisSinceEpoch(0),
